+++ src/SlyTwitter/twitter.py
@@ -4,13 +4,11 @@ from typing import Any
 from SlyAPI import *
 
 from .twitter_upload import Media, TwitterUpload
-from .common import make_with_self
 
 RE_TWEET_LINK = re.compile(r'https://twitter\.com/(?P<user>[a-z0-9_]+)/status/(?P<tweet_id>\d+)', re.IGNORECASE)
 RE_USER_LINK = re.compile(r'https://twitter\.com/(?P<user>[a-z0-9_]+)', re.IGNORECASE)
 
-
-class User(APIObj['Twitter']):
+class User:
     id: int
     at: str
     display_name: str
@@ -24,8 +22,7 @@ class User(APIObj['Twitter']):
     created_at: datetime|None = None
     profile_image: str|None = None
 
-    def __init__(self, source: dict[str, Any], service: 'Twitter'):
-        super().__init__(service)
+    def __init__(self, source: dict[str, Any]):
         match source:
             case int():
                 self.id = source
@@ -69,20 +66,19 @@ class User(APIObj['Twitter']):
     def __str__(self):
         return F'@{self.at}'
 
-class Following(APIObj['Twitter']):
+class Following:
     a: User
     b: User
     is_mutual: bool
     a_follows_b: bool
     b_follows_a: bool
 
-    def __init__(self, source: dict[str, Any], service: 'Twitter'):
-        super().__init__(service)
+    def __init__(self, source: dict[str, Any]):
         self.a_follows_b = source['relationship']['source']['following']
         self.b_follows_a = source['relationship']['target']['following']
         self.mutual = self.a_follows_b and self.b_follows_a
-        self.a = User(source['relationship']['source'], service)
-        self.b = User(source['relationship']['target'], service)
+        self.a = User(source['relationship']['source'])
+        self.b = User(source['relationship']['target'])
 
     def __str__(self) -> str:
         if self.mutual:
@@ -96,37 +92,34 @@ class Following(APIObj['Twitter']):
         return F"{self.a} {rel_str} {self.b}"
 
 
-class Tweet(APIObj['Twitter']):
+class Tweet:
     id: int
-    author_username: str # twitter user @
+    author_at: str # twitter user @
     body: str
 
-    def __init__(self, source: int | str | dict[str, Any], service: 'Twitter'):
-        super().__init__(service)
+    def __init__(self, source: int | str | dict[str, Any]):
         match source:
             case int():
                 self.id = source
             case str():
                 if not (m := RE_TWEET_LINK.match(source)):
                     raise ValueError('Cannot create Tweet without ID, link to tweet, or dict representation')
-                self.author_username = m['user']
+                self.author_at = m['user']
                 self.id = int(m['tweet_id'])
             case { 'id': id_, 'extended_tweet': { 'full_text': text }, 'user': { 'screen_name': user } }:
                 self.id = id_
                 self.body = text
-                self.author_username = user
+                self.author_at = user
             case { 'id': id_, 'text': text, 'user': { 'screen_name': user } }:
                 self.id = id_
                 self.body = text
-                self.author_username = user
+                self.author_at = user
             case _:
                 raise TypeError(F"{source} is not a valid source for Tweet")
 
     def link(self) -> str:
-        return F"https://twitter.com/{self.author_username}/status/{self.id}"
+        return F"https://twitter.com/{self.author_at}/status/{self.id}"
 
-    async def delete(self):
-        await self._service.delete(self)
     
 def get_tweet_id(tweet: Tweet | int | str) -> int:
     match tweet:
@@ -141,20 +134,18 @@ def get_tweet_id(tweet: Tweet | int | str) -> int:
 
 
 class Twitter(WebAPI):
+    '''Twitter V1.1 API'''
     base_url = 'https://api.twitter.com/1.1'
     _upload_api: TwitterUpload
     
-    async def __init__(self, auth: str | OAuth1User):
-        if isinstance(auth, str):
-            auth = OAuth1User(auth)
+    def __init__(self, auth: OAuth1):
 
-        await super().__init__(auth)
-        self._upload_api = await TwitterUpload(auth)
+        super().__init__(auth)
+        self._upload_api = TwitterUpload(auth)
 
     def get_full_url(self, path: str) -> str:
         return super().get_full_url(path) +'.json'
 
-    @make_with_self(Tweet)
     async def tweet(self, body: str, media: list[Media] | str | tuple[bytes, str] | None = None):
         """
             Post a tweet.
@@ -169,27 +160,26 @@ class Twitter(WebAPI):
             media = [await self._upload_api.upload(media)]
         if media:
             data |= { 'media_ids': ','.join(str(m.id) for m in media) }
-        return await self.post_json( '/statuses/update',
+        return Tweet(await self.post_form( '/statuses/update',
             data = data
-        )
+        ))
 
-    @make_with_self(Following)
     async def check_follow(self, a: User | str, b: User | str):
         """ Get the relationship between two users. """
         if isinstance(a, User): a = a.at
         if isinstance(b, User): b = b.at
-        return await self.get_json( '/friendships/show', {
+        return Following(await self.get_json( '/friendships/show', {
             'source_screen_name': a,
             'target_screen_name': b
-        })
+        }))
 
     async def delete(self, tweet: Tweet | int | str):
-        id = get_tweet_id(tweet)
-        await self.post_json(F'/statuses/destroy/{id}')
+        tweet_id = get_tweet_id(tweet)
+        await self.post_json(F'/statuses/destroy/{tweet_id}')
 
     async def retweet(self, tweet: Tweet | int | str):
-        id = get_tweet_id(tweet)
-        await self.post_json(F'/statuses/retweet/{id}')
+        tweet_id = get_tweet_id(tweet)
+        await self.post_json(F'/statuses/retweet/{tweet_id}')
 
     async def quote_tweet(self, body: str, quoting: Tweet | str, media: list[Media] | str | tuple[bytes, str] | None = None) -> Tweet:
         if isinstance(quoting, Tweet):
@@ -207,3 +197,7 @@ class Twitter(WebAPI):
               - a bytes-like obj a tupled with a file extension
         """
         return await self._upload_api.upload(file_)
+    
+    async def add_alt_text(self, media: Media, text: str):
+        """ Add alt text to a media file. """
+        await self._upload_api.add_alt_text(media, text)