+++ test/test_readme.py
@@ -3,10 +3,15 @@ import asyncio
 
 import aiohttp
 from SlyTwitter import *
+from SlyTwitter.twitter import OAuth1, OAuth1App, OAuth1User
+
+app = OAuth1App.from_json_file('test/sly_test_app.json')
+user = OAuth1User.from_json_file('test/user.json')
+auth = OAuth1(app, user)
 
 async def test_readme():
 
-    twitter = await Twitter('test/user.json')
+    twitter = Twitter(auth)
 
     # tweet = await twitter.tweet('Hello, world!')
     follow = await twitter.check_follow('dunkyl_', 'TechConnectify')
@@ -17,12 +22,12 @@ async def test_readme():
 @pytest.mark.skip(reason="effectual")
 async def test_upload_tweet_delete():
 
-    twitter = await Twitter('test/user.json')
+    twitter = Twitter(auth)
 
     # post a tweet with an image
 
     media = await twitter.upload_media('test/test.jpg')
-    await media.add_alt_text('A test image.')
+    await twitter.add_alt_text(media, 'A test image.')
     tweet = await twitter.tweet('Hello, world!', [media])
 
     print(tweet)
@@ -31,7 +36,7 @@ async def test_upload_tweet_delete():
 
     # delete it and make sure its gone
 
-    await tweet.delete()
+    await twitter.delete(tweet)
 
     await asyncio.sleep(10)