+++ test/test_readme_v2.py
@@ -1,10 +1,15 @@
 import pytest
 from SlyTwitter.twitter_v2 import TwitterV2 as Twitter
+from SlyTwitter.twitter_v2 import OAuth2, OAuth2App, OAuth2User
 
 @pytest.mark.skip(reason="v2")
 async def test_readme_v2():
 
-    twitter = await Twitter('test/twauth2.json', 'test/user22.json')
+    app = OAuth2App.from_json_file('test/twauth2.json')
+    user = OAuth2User.from_json_file('test/user22.json')
+    auth = OAuth2(app, user)
+
+    twitter = Twitter(auth)
 
     me = await twitter.user()